some test added

+++ src/main/java/efimovta/store/dao/SaleDAO.java
@@ -17,7 +17,7 @@ public interface SaleDAO extends GenericDAO<Sale> {
      * @param id client id
      * @return list of {@link Sale} objects founded by client id
      */
-    public List<Sale> findBySaleId(int id) throws DAOException;
+    public List<Sale> findBySaleId(long id) throws DAOException;
 
     /**
      * Finds the {@link Sale} objects by client id
@@ -25,7 +25,7 @@ public interface SaleDAO extends GenericDAO<Sale> {
      * @param id client id
      * @return list of {@link Sale} objects founded by client id
      */
-    public List<Sale> findByClientId(int id) throws DAOException;
+    public List<Sale> findByClientId(long id) throws DAOException;
 
     /**
      * Finds the {@link Sale} objects by device id
@@ -33,7 +33,7 @@ public interface SaleDAO extends GenericDAO<Sale> {
      * @param id device id
      * @return list of {@link Sale} objects founded by device id
      */
-    public List<Sale> findByDeviceId(int id) throws DAOException;
+    public List<Sale> findByDeviceId(long id) throws DAOException;
 
     /**
      * Finds the {@link Sale} objects by sale date
